Add IndexSpec.
